token-harvest-attack/requirements.txt
aiohttp==3.8.1
PyGithub